Add parsing of math blocks
Without a library like Katex, this is just rendered as a blue span of
text, but we can easily fix that soon enough. The tricky part here is
that math blocks can contain other markdown characters. The way the
parser is architected seems way too complicated for our use case, so we
might want to slim things down.