Create RemoteSea interface
This will be used to wrap our HTTP communication with the backend. Right
now it just has a method to lookup the string contents of a Bubble given
its ID. The logic behind the string contents is that you don't want to
have to have parsing logic in the implementation of the interface.